% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanHippocampus2024-package.R
\docType{package}
\name{humanHippocampus2024-package}
\alias{humanHippocampus2024}
\alias{humanHippocampus2024-package}
\alias{human-hippocampus}
\title{Get access to the spatial HPC project data}
\description{
This is an ExperimentHub package that provides access to the
data generated and analyzed in the
\href{https://github.com/LieberInstitute/spatial_hpc}{humanHippocampus2024}
project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/spatial_hpc}
  \item \url{https://christinehou11.github.io/humanHippocampus2024/}
  \item Report bugs at \url{https://support.bioconductor.org/t/humanHippocampus2024}
}

}
\author{
Christine Hou
}
\keyword{internal}
