% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialHPC-package.R
\docType{package}
\name{spatialHPC-package}
\alias{spatialHPC}
\alias{spatialHPC-package}
\alias{human-hippocampus}
\title{spatialHPC: Access to spatialHPC project data}
\description{
What the package does (one paragraph).

This is an ExperimentHub package that provides access to the
data generated and analyzed in the
\href{https://github.com/LieberInstitute/spatial_hpc}{spatialHPC} project.
License: Artistic-2.0
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/spatial_hpc}
  \item \url{https://christinehou11.github.io/spatialHPC/}
  \item Report bugs at \url{https://support.bioconductor.org/t/spatialHPC}
}

}
\author{
\strong{Maintainer}: Christine Hou \email{chris2018hou@gmail.com} (\href{https://orcid.org/0009-0001-5350-0629}{ORCID})


Christine Hou
}
\keyword{internal}
