% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanHippocampus2024-package.R
\docType{package}
\name{humanHippocampus2024-package}
\alias{humanHippocampus2024}
\alias{humanHippocampus2024-package}
\alias{human-hippocampus}
\title{humanHippocampus2024: Access to SRT and snRNA-seq data from spatial_HPC project}
\description{
This is an ExperimentHub package that helps to access the spatially-resolved transcriptomics and single-nucleus RNA sequencing data. The datasets are generated from adjacent tissue sections of the anterior human hippocampus across ten adult neurotypical donors. The datasets are based on [spatial_hpc](https://github.com/LieberInstitute/spatial_hpc) project by Lieber Institute for Brain Development (LIBD) researchers and collaborators.

This is an ExperimentHub package that provides access to the
data generated and analyzed in the
\href{https://github.com/LieberInstitute/spatial_hpc}{humanHippocampus2024} project.
License: Artistic-2.0
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/spatial_hpc}
  \item \url{https://christinehou11.github.io/humanHippocampus2024/}
  \item Report bugs at \url{https://support.bioconductor.org/t/humanHippocampus2024}
}

}
\author{
\strong{Maintainer}: Christine Hou \email{chris2018hou@gmail.com} (\href{https://orcid.org/0009-0001-5350-0629}{ORCID})


Christine Hou
}
\keyword{internal}
